refactor(monitoring): TSDoc 작성 및 함수 분리
